backend/requirements.txt
Django==6.0.4
python-dotenv==1.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.5
Pillow==11.1.0